refactor: mcp server index into 4 files
