+++ mcp-server/src/stdioServer.ts
@@ -1,0 +1,6 @@
+import type { McpServer } from '@modelcontextprotocol/sdk/server/mcp.js';
+import { StdioServerTransport } from '@modelcontextprotocol/sdk/server/stdio.js';
+
+export async function startStdioServer(factory: () => McpServer): Promise<void> {
+  await factory().connect(new StdioServerTransport());
+}
